Allow only the logged in user to manage tasks

+++ task/views.py
@@ -29,7 +29,10 @@ def todolist(request):
 @login_required()
 def delete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.delete() 
+    if task.manage == request.user:
+        task.delete() 
+    else:
+        messages.error(request, ("Access Denied, You Are Not Allowed!"))
     return redirect('todolist')
 
 
@@ -51,9 +54,11 @@ def edit_task(request, task_id):
 @login_required()
 def complete_task(request, task_id):
     task = TaskList.objects.get(pk=task_id)
-    task.done = True    
-    task.save()
-
+    if task.manage == request.user:
+        task.done = True    
+        task.save()
+    else:
+        messages.error(request, ("Access Denied, You Are Not Allowed!"))
     return redirect('todolist')
